+++ src/resources/agents/index.ts
@@ -19,7 +19,6 @@ export {
   type RequiresApprovalToolRule,
   type TerminalToolRule,
   type TextResponseFormat,
-  type AgentListResponse,
   type AgentDeleteResponse,
   type AgentCountResponse,
   type AgentExportFileResponse,
@@ -30,17 +29,18 @@ export {
   type AgentListParams,
   type AgentExportFileParams,
   type AgentImportFileParams,
+  type AgentStatesArrayPage,
 } from './agents';
 export {
   Blocks,
   type Block,
   type BlockUpdate,
-  type BlockListResponse,
   type BlockRetrieveParams,
   type BlockUpdateParams,
   type BlockListParams,
   type BlockAttachParams,
   type BlockDetachParams,
+  type BlocksArrayPage,
 } from './blocks';
 export {
   Files,
@@ -51,6 +51,7 @@ export {
   type FileListParams,
   type FileCloseParams,
   type FileOpenParams,
+  type FileListResponsesNextFilesPage,
 } from './files';
 export {
   Folders,
@@ -58,8 +59,9 @@ export {
   type FolderListParams,
   type FolderAttachParams,
   type FolderDetachParams,
+  type FolderListResponsesArrayPage,
 } from './folders';
-export { Groups, type GroupListResponse, type GroupListParams } from './groups';
+export { Groups, type GroupListParams } from './groups';
 export {
   Messages,
   type ApprovalCreate,
@@ -98,7 +100,6 @@ export {
   type UpdateUserMessage,
   type UserMessage,
   type MessageUpdateResponse,
-  type MessageListResponse,
   type MessageCancelResponse,
   type MessageStreamResponse,
   type MessageUpdateParams,
@@ -108,10 +109,12 @@ export {
   type MessageSendParams,
   type MessageSendAsyncParams,
   type MessageStreamParams,
+  type LettaMessageUnionsArrayPage,
+  type RunsArrayPage,
+  type MessagesObjectPage,
 } from './messages';
 export {
   Tools,
-  type ToolListResponse,
   type ToolListParams,
   type ToolAttachParams,
   type ToolDetachParams,