+++ src/resources/agents/tools.ts
@@ -2,8 +2,10 @@
 
 import { APIResource } from '../../core/resource';
 import * as ToolsAPI from '../tools';
+import { ToolsArrayPage } from '../tools';
 import * as AgentsAPI from './agents';
 import { APIPromise } from '../../core/api-promise';
+import { ArrayPage, type ArrayPageParams, PagePromise } from '../../core/pagination';
 import { RequestOptions } from '../../internal/request-options';
 import { path } from '../../internal/utils/path';
 
@@ -15,8 +17,11 @@ export class Tools extends APIResource {
     agentID: string,
     query: ToolListParams | null | undefined = {},
     options?: RequestOptions,
-  ): APIPromise<ToolListResponse> {
-    return this._client.get(path`/v1/agents/${agentID}/tools`, { query, ...options });
+  ): PagePromise<ToolsArrayPage, ToolsAPI.Tool> {
+    return this._client.getAPIList(path`/v1/agents/${agentID}/tools`, ArrayPage<ToolsAPI.Tool>, {
+      query,
+      ...options,
+    });
   }
 
   /**
@@ -26,7 +31,7 @@ export class Tools extends APIResource {
     toolID: string,
     params: ToolAttachParams,
     options?: RequestOptions,
-  ): APIPromise<AgentsAPI.AgentState> {
+  ): APIPromise<AgentsAPI.AgentState | null> {
     const { agent_id } = params;
     return this._client.patch(path`/v1/agents/${agent_id}/tools/attach/${toolID}`, options);
   }
@@ -38,7 +43,7 @@ export class Tools extends APIResource {
     toolID: string,
     params: ToolDetachParams,
     options?: RequestOptions,
-  ): APIPromise<AgentsAPI.AgentState> {
+  ): APIPromise<AgentsAPI.AgentState | null> {
     const { agent_id } = params;
     return this._client.patch(path`/v1/agents/${agent_id}/tools/detach/${toolID}`, options);
   }
@@ -53,7 +58,7 @@ export class Tools extends APIResource {
     toolName: string,
     params: ToolUpdateApprovalParams,
     options?: RequestOptions,
-  ): APIPromise<AgentsAPI.AgentState> {
+  ): APIPromise<AgentsAPI.AgentState | null> {
     const { agent_id, query_requires_approval, ...body } = params;
     return this._client.patch(path`/v1/agents/${agent_id}/tools/approval/${toolName}`, {
       query: { requires_approval: query_requires_approval },
@@ -63,37 +68,7 @@ export class Tools extends APIResource {
   }
 }
 
-export type ToolListResponse = Array<ToolsAPI.Tool>;
-
-export interface ToolListParams {
-  /**
-   * Tool ID cursor for pagination. Returns tools that come after this tool ID in the
-   * specified sort order
-   */
-  after?: string | null;
-
-  /**
-   * Tool ID cursor for pagination. Returns tools that come before this tool ID in
-   * the specified sort order
-   */
-  before?: string | null;
-
-  /**
-   * Maximum number of tools to return
-   */
-  limit?: number | null;
-
-  /**
-   * Sort order for tools by creation time. 'asc' for oldest first, 'desc' for newest
-   * first
-   */
-  order?: 'asc' | 'desc';
-
-  /**
-   * Field to sort by
-   */
-  order_by?: 'created_at';
-}
+export interface ToolListParams extends ArrayPageParams {}
 
 export interface ToolAttachParams {
   /**
@@ -128,10 +103,11 @@ export interface ToolUpdateApprovalParams {
 
 export declare namespace Tools {
   export {
-    type ToolListResponse as ToolListResponse,
     type ToolListParams as ToolListParams,
     type ToolAttachParams as ToolAttachParams,
     type ToolDetachParams as ToolDetachParams,
     type ToolUpdateApprovalParams as ToolUpdateApprovalParams,
   };
 }
+
+export { type ToolsArrayPage };